[refactor/#217] :feature:mypage 내 패키지 구조 수정
