+++ feature/mypage/src/main/java/com/napzak/market/mypage/mypage/MyPageViewModel.kt
@@ -1,8 +1,8 @@
-package com.napzak.market.mypage.model
+package com.napzak.market.mypage.mypage
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.napzak.market.mypage.state.MyPageUiState
+import com.napzak.market.mypage.mypage.state.MyPageUiState
 import com.napzak.market.store.repository.StoreRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
