+++ feature/mypage/src/main/java/com/napzak/market/mypage/mypage/type/MyPageMenu.kt
@@ -1,19 +1,19 @@
-package com.napzak.market.mypage
+package com.napzak.market.mypage.mypage.type
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
-import com.napzak.market.designsystem.R.drawable.ic_sales_36
+import com.napzak.market.designsystem.R.drawable.ic_favorite_36
+import com.napzak.market.designsystem.R.drawable.ic_help_36
 import com.napzak.market.designsystem.R.drawable.ic_purchase_36
 import com.napzak.market.designsystem.R.drawable.ic_recent_36
-import com.napzak.market.designsystem.R.drawable.ic_favorite_36
+import com.napzak.market.designsystem.R.drawable.ic_sales_36
 import com.napzak.market.designsystem.R.drawable.ic_settings_36
-import com.napzak.market.designsystem.R.drawable.ic_help_36
-import com.napzak.market.feature.mypage.R.string.mypage_sales_history
+import com.napzak.market.feature.mypage.R.string.mypage_favorites
+import com.napzak.market.feature.mypage.R.string.mypage_help_center
 import com.napzak.market.feature.mypage.R.string.mypage_purchase_history
 import com.napzak.market.feature.mypage.R.string.mypage_recently_viewed
-import com.napzak.market.feature.mypage.R.string.mypage_favorites
+import com.napzak.market.feature.mypage.R.string.mypage_sales_history
 import com.napzak.market.feature.mypage.R.string.mypage_settings
-import com.napzak.market.feature.mypage.R.string.mypage_help_center
 
 enum class MyPageMenu(
     @StringRes val titleRes: Int,
